rag-knowledge-assistant/requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu
python-dotenv
pymupdf